Megaturbodisco]

+++ swont_ide/Core/LogicLayer/logic.h
@@ -4,7 +4,7 @@
 
 #include <stdint.h>
 
-uint8_t logicAPICallLijn(uint8_t x1p, uint8_t y1p, uint8_t x2p, uint8_t y2p, const char *color, uint8_t siz);
+uint8_t logicAPICallLijn(uint16_t x1p, uint16_t y1p, uint16_t x2p, uint16_t y2p, const char *color, uint8_t siz);
 uint8_t logicAPICallfill(const char *color);
 uint8_t logicAPICallRechthoek(uint8_t x1p, uint8_t y1p, uint8_t x2p, uint8_t y2p, const char *color, uint8_t filled);
 uint8_t logicAPICallcirkel(uint8_t xp, uint8_t yp, uint8_t siz, const char *color);
